+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The `pycldf` package adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
-## unreleased
+## [1.40.0] - 2024-12-13
 
 - Support certain GitHub URLs as dataset locators.
 - Pragmatic support for large media files.